requirements (2).txt
tensorflow
streamlit
opencv-python
Pillow
scikit-image
matplotlib
